moved react components to separate folder, moved javascript function files to scripts folder
